Fix solution to LeetCode problem #219
The original solution was incorrect. The solution compared the original
indices of adjacent numbers after sorting them. However, as mentioned in
the comments of the solution, the sort is not stable and, in theory, it
is possible for a number to be inserted in between 2 equal numbers whose
original indices are within k of each other, but whose original indices
are not within k of the original index of the inserted element.
Thereafter, iterating through the adjacent elements would errouneously
return false.

In addition, the runtime of the current solution is better than the
original solution.

Original runtime complexity: O(n*logn)
Current runtime complexity: O(n)

Where n is the size of the input array.