Improve solution to LeetCode problem #441
In our original approach, we determine the number of levels by
iteratively going row by row to see if we had enough to coins to make
that row. The number of rows we create is k where k is the largest value
for which the sum from 1 to k (i.e., the sum of all coins in the k rows)
is less than or equal to n. Based on our understanding of triangle
numbers, such a value is equal to (or 1 less than) the value m such that
n = (m * (m + 1)) / 2. Using the quadratic formula and substituting for
k we get that k is O(sqrt(n)).

On the other hand, this knowledge can also inform an improved approach.
We can set bounds for the number of rows based on the maximum and
minimum triangle numbers that fit within the constraints (i.e., 1 <= n
<= 2^31 -1). Then, we can perform a binary search where we check, in
constant time, with our formula for the the sum of integers 1 to m, the
number of coins required to make m rows. Thereafter, we can update the
bounds. Each iteration halves the search space until we find the
appropriate number of rows.

Technically, this algorithm can be said to run in constant time since we
are dealing with fixed bounds and halving these bounds predictably.
However, these bounds emerge from an understanding of the constraints.
The upper bound comes from the largest possible value of our input n and
is in proportion to its square root. Thus, we can argue that the runtime
is Ω(log(sqrt(n))) for all values of n and Θ(log(sqrt(n)) for the
largest value of n.