Solve LeetCode problem #234
Given the head of a singly linked list, return true if it is a
palindrome and false otherwise.

The approach in this solution works by reversing the second half of the
linked list and then comparing the 2 halves with one another. This
approach satisfies the constraints of O(n) time and O(1) space.

However, since it modifies the underlying data passed to it to determine
whether the list is a palindrome, the side effects are undesirable. An
alternative approach would be to use a vector/stack to store the first
half of the list's elements and then compare them with the second half.
Such an approach would have the same runtime complexity, but the space
complexity would be O(n).