Update solution to LeetCode problem #977
The original approach involved searching for the first non-negative
value's index and then reversing the list of negative numbers on the
left of this index. This effectively sorted it into ascending order of
absolute value. Then, two pointers were used to compare the smallest (in
absolute value) negative value with the smallest non-negative value
until all values of the original array had been collated into a new
array in order of absolute value. Finally, a last pass-through was done
to square all the numbers in the new array of collated values.

The current approach initializes the output array with the correct size
(i.e., the same size as the input array). Then, the output is built
backwards (i.e., from the largest squared value to the smallest). To do
this, two pointers are used to track the smallest and largest values.
One of the smallest and largest values is guaranteed to have the largest
absolute value of all the values in array. The output is generated by
squaring that value and adding it to the back of the output array. The
indices for the smallest and largest value are updated accordingly.

Original Runtime Complexity: O(n)
Current Runtime Complexity: O(n)

Original Space Complexity: O(n)
Current Space Complexity: O(n)