Solve LeetCode problem #26
Given an array sorted in non-decreasing order, remove the duplicates
in-place such that each unique element appears only once. The relative
order of the elements should be kept the same. Return the number of
unique elements in the input array.

Note: The input array only has to be modified so that it starts with all
the unique elements in their correct order. The remainder of the array
or its size does not need to be modified.