Improve solution to LeetCode problem #160
The original solution uses an unordered set to store all nodes in the
first linked list and traverses the second linked list to see if any of
its nodes are in the first linked list.

The present solution calculates the size of the linked lists and then
compares the elements at the head of each list. If the values are the
same, the current head of a list is returned. However, if they are
different, we advance the pointer of the larger of the remaining lists
and update its size. In the event that the sizes are the same, but we
did not have a match, we will advance the pointer and update the size of
any one list (i.e., the second list in our case).

This new approach works because it always compares elements that are the
same distance apart from their respective ends. The lists may intersect
only when their remaining sizes are the same since the remainder of the
list would be the same.

Original Runtime Complexity: O(n + m)
Original Space Complexity: O(n)

Current Runtime Complexity: O(max(n, m))
Current Space Complexity: O(1)

The runtime of max(n, m) is in effect the same as n + m because
whichever list is longer will dominate the sum and that means the
runtime of n + m is capped at 2*max(n, m),  which is on the order of
max(m, n). On the other hand, max(n, m) < n + m so any algorithm that
runs in max(n, m) is on the order of n + m.