Update solution to LeetCode problem #1888
There are 2 possible alternating binary strings for any given length.
One starting with 0 and the other starting with 1. The initial approach
explicitly tracks the number of flips required for both states; however,
the number of flips to achieve one state is the complement of the number
of flips to achieve the second state.

In particular, the relationship is as follows:

(number of flips to achieve 1010...) = (
  (size of binary string) - (number of flips to achieve 0101...)
)

All flips that are required for (0101...) are not required for (1010...)
because they represent bits that are the opposite of what is expected in
(0101...), but exactly what is in (1010...). On the other hand, the
remaining bits from the total number of bits (i.e., the size of the
binary string) match the expected bits from (0101...); however, these
bits are the opposite of what is required in (1010...) and must be
flipped.

So, in this approach the solution is simplified by only tracking how
many flips are required to achieve one of the binary strings and
determining the number of flips for the other binary string based on
this value and the relationship above.

Runtime Complexity: O(n)
Space Complexity: O(1)