Improve solution to LeetCode problem #703
The original implementation initializes a min-heap and inserts all
provided numbers to the heap before removing all but k elements. The
insertion of elements after the kth element takes greater than
O(log(k)). So, to ensure that every operation of insertion in the
constructor is bound by O(log(k)), we perform a check on the size of the
heap after insertion so that our heap never contains more than k + 1
elements at a given time.

Original Runtime Complexity: O(a*log(a) + b*log(k))
Original Space Complexity: O(n)

Current Runtime Complexity: O(n*log(k))
Current Runtime Complexity: O(k)

Where a represents the size of the array passed to the constructor and b
represents the number of calls to the add method.

Note: n is the total number of elements in the entire stream.
n = a + b