Solve LeetCode problem #989
Given the array-form of an integer n and an integer k, return the
array-form of the integer n + k.

Note: An array-form an integer is an array of digits representing a
decimal number where the first element corresponds to the largest digit,
the second element corresponds to the second largest digit, and so on.