Solve LeetCode problem #80
Given an integer array sorted in non-decreasing order, remove the
duplicates of the array in-place such that unique elements appear at
most twice. The relative order of the elements should be kept the same.

The array does not need to resized; if k elements remain after the
removals, then the first k elements of the array should be correct and
the rest of the array is not relevant.

Return the integer value k (i.e., size of the array after removals).

Runtime Complexity: O(n)
Space Complexity: O(1)