Update solution to LeetCode problem #27
The original approach starts from the end of the array and swaps the
values to be removed to the end. On the other hand, the current approach
starts at the beginning of the array and adds valid elements as it goes.

The benefit of the current approach is that it maintains the relative
order of elements from the original array.

Original Runtime Complexity: O(n)
Original Space Complexity: O(1)

Current Runtime Complexity: O(n)
Current Space Complexity: O(1)