Solve LeetCode problem #929
Given an array of strings representing email addresses, return the
number of unique email addresses.

Note: An email consists of a local name and a domain name delimited by
the '@' character. The '.' character is ignored in the local name. Any
characters after, and including, the first instance of the '+' character
in the local name are ignored.