Solve LeetCode problem #496
Given an integer array of unique elements called nums1 and another array
which contains a subset of the elements from nums1 called nums2, return
an array that contains the next greater element for every element in
nums1.

The next greater element is defined as the element from nums1 that
occurs sequentially after the specified number. If such a number does
not exist, the value -1 is used.

Note: The length of the nums1 array is denoted as n below.

Runtime Complexity: O(n)
Space Complexity: O(n)