Solve LeetCode problem #735
Given an integer array representing asteroids with the magnitude as the
size and the sign as the direction (i.e., negative is left, positive is
right), return an array with the asteroids after collisions.

Note: If 2 asteroids of the same size collide, they are both destroyed.
If 2 asteroids of different sizes collide, the smaller asteroid is
destroyed.