Solve LeetCode problem #881
Given a positive integer array representing the weights of people and an
integer limit, if a boat can carry at most 2 people with a maximum
capacity of limit, return the minimum number of boats required to carry
all people.

Note: Every person's weight is less than or equal to limit.