Update solution to LeetCode problem #876
The original solution calculates the size of the linked list and then
determines the index of the middle node to traverse to it. The current
solution uses a sentinel node and 2 pointers (i.e., slow that traverses
1 node at a time and fast that traverses 2 nodes at a time). The slow
pointer is returned when the fast pointer passes the end of the list.

Original Runtime Complexity: O(n)
Original Space Complexity: O(1)

Current Runtime Complexity: O(n)
Current Space Complexity: O(1)